Formula for repeatable ipfs builds.
Includes some bash hacks to loop over cross-compiling for a range of architectures and OSes.  The input, however, assumes linux and amd64.  (Scripting a windows build is a whole different ballgame.  Fortunately, it seems to be popularly acceptable to drive builds out of linux cross-compiles; tor, gitian, and others take the same approach.)